Replace double with BigDecimal for money calculation

+++ src/main/java/com/example/rewardsprogram/api/model/PurchaseData.java
@@ -2,14 +2,15 @@ package com.example.rewardsprogram.api.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class PurchaseData {
-
     private Long id;
     private String name;
-    private double amount;
+    private BigDecimal amount;
 }
