+++ src/main/java/com/example/rewardsprogram/domain/Transaction.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -20,7 +21,7 @@ public class Transaction {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private double amount;
+    private BigDecimal amount;
 
     private LocalDate date;
 
@@ -28,6 +29,4 @@ public class Transaction {
     @JoinColumn(name = "customer_id", nullable = false)
     @JsonBackReference
     private Customer customer;
-
-    // getters and setters omitted for brevity
 }
